pyproject.toml
# pyproject.toml
[tool.poetry]
name = "horse-quant-project"
version = "0.1.0"
description = "Horse Racing ML pipeline skeleton (bootstrap stage)."
authors = ["Dylan McGuinness <dylanmcguinness0@gmail.com>"]
readme = "README.md"
packages = [{ include = "hqp", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
PyYAML = "^6.0.2"
pydantic = "<2"
pandas = "^2.3.2"
pyarrow = "^21.0.0"
pandera = "^0.26.1"
numpy = "^2.3.3"
typer = "^0.12.5"
click = "==8.1.7"
rich = ">=13"
tzdata = "^2025.1"
scikit-learn = "^1.7.2"
lightgbm = "^4.6.0"
matplotlib = "^3.9.0"  # plotting backend
tabulate = "^0.9.0"
tqdm = "^4.67.1"

[tool.poetry.scripts]
hqp = "hqp.cli:app"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
ruff = "^0.6.5"
mypy = "^1.11.2"
pytest = "^8.3.2"
pre-commit = "^3.8.0"
types-PyYAML = "^6.0.12.20240808"
pandas-stubs = "^2.3.2.250827"
pytest-cov = "^7.0.0"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP"]
ignore = ["E203"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
disallow_any_generics = false
files = ["src"]
namespace_packages = true
explicit_package_bases = true

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
